apu serialize

+++ core/gb/apu/io.go
@@ -44,49 +44,49 @@ func (a *apu) Write(addr uint16, val uint8) {
 		if a.ch1.sweep.enabled && a.ch1.sweep.negate && !negate {
 			a.ch1.enabled = false // 下降スイープ中に上昇スイープを設定すると消音される
 		}
-		a.ch1.sweep.shift = int(val & 0b111)
+		a.ch1.sweep.shift = int8(val & 0b111)
 		a.ch1.sweep.negate = negate
-		a.ch1.sweep.interval = int(val>>4) & 0b111
+		a.ch1.sweep.interval = int8(val>>4) & 0b111
 
 	case 0xFF11:
-		a.ch1.duty = int(val >> 6)
-		a.ch1.length = 64 - int(val&0b11_1111)
+		a.ch1.duty = (val >> 6)
+		a.ch1.length = 64 - int32(val&0b11_1111)
 
 	case 0xFF12:
-		a.ch1.envelope.initialVolume = int(val>>4) & 0b1111
+		a.ch1.envelope.initialVolume = int32(val>>4) & 0b1111
 		a.ch1.envelope.direction = util.Bit(val, 3)
-		a.ch1.envelope.speed = int(val & 0b111)
+		a.ch1.envelope.speed = int32(val & 0b111)
 		if !a.ch1.dacEnable() {
 			a.ch1.enabled = false
 		}
 
 	case 0xFF13:
-		a.ch1.period = (a.ch1.period & 0xFF00) | int(val)
+		a.ch1.period = (a.ch1.period & 0xFF00) | int32(val)
 
 	case 0xFF14:
-		a.ch1.period = (a.ch1.period & 0x00FF) | (int(val&0b111) << 8)
+		a.ch1.period = (a.ch1.period & 0x00FF) | (int32(val&0b111) << 8)
 		a.ch1.stop = util.Bit(val, 6)
 		if util.Bit(val, 7) { // キーオン(音が鳴り始める)
 			a.ch1.tryRestart()
 		}
 
 	case 0xFF16:
-		a.ch2.duty = int(val >> 6)
-		a.ch2.length = 64 - int(val&0b11_1111)
+		a.ch2.duty = (val >> 6)
+		a.ch2.length = 64 - int32(val&0b11_1111)
 
 	case 0xFF17:
-		a.ch2.envelope.initialVolume = int(val>>4) & 0b1111
+		a.ch2.envelope.initialVolume = int32(val>>4) & 0b1111
 		a.ch2.envelope.direction = util.Bit(val, 3)
-		a.ch2.envelope.speed = int(val & 0b111)
+		a.ch2.envelope.speed = int32(val & 0b111)
 		if !a.ch2.dacEnable() {
 			a.ch2.enabled = false
 		}
 
 	case 0xFF18:
-		a.ch2.period = (a.ch2.period & 0xFF00) | int(val)
+		a.ch2.period = (a.ch2.period & 0xFF00) | int32(val)
 
 	case 0xFF19:
-		a.ch2.period = (a.ch2.period & 0x00FF) | (int(val&0b111) << 8)
+		a.ch2.period = (a.ch2.period & 0x00FF) | (int32(val&0b111) << 8)
 		a.ch2.stop = util.Bit(val, 6)
 		if util.Bit(val, 7) { // キーオン(音が鳴り始める)
 			a.ch2.tryRestart()
@@ -98,21 +98,21 @@ func (a *apu) Write(addr uint16, val uint8) {
 			a.ch3.enabled = false
 		}
 		if a.model == MODEL_GBA {
-			a.ch3.mode = int(val>>5) & 0b1
-			a.ch3.bank = int(val>>6) & 0b1
+			a.ch3.mode = uint8(val>>5) & 0b1
+			a.ch3.bank = uint8(val>>6) & 0b1
 		}
 	case 0xFF1B:
-		a.ch3.length = 256 - int(val)
+		a.ch3.length = 256 - int32(val)
 	case 0xFF1C:
-		a.ch3.volume = [4]int{4, 0, 1, 2}[int(val>>5)&0b11] // 波形は最大15なので4左シフトすれば0%
+		a.ch3.volume = [4]uint8{4, 0, 1, 2}[int(val>>5)&0b11] // 波形は最大15なので4左シフトすれば0%
 	case 0xFF1D:
 		a.ch3.period &= 0b111_0000_0000
-		a.ch3.period |= int(val)
+		a.ch3.period |= int32(val)
 		a.ch3.freqCounter = a.ch3.windowStepCycle()
 	case 0xFF1E:
 		a.ch3.stop = util.Bit(val, 6)
 		a.ch3.period &= 0b000_1111_1111
-		a.ch3.period |= int(val&0b111) << 8
+		a.ch3.period |= int32(val&0b111) << 8
 		a.ch3.freqCounter = a.ch3.windowStepCycle()
 		if util.Bit(val, 7) { // キーオン(音が鳴り始める)
 			a.ch3.enabled = a.ch3.dacEnable
@@ -130,17 +130,17 @@ func (a *apu) Write(addr uint16, val uint8) {
 		}
 
 	case 0xFF20:
-		a.ch4.length = 64 - int(val&0b11_1111)
+		a.ch4.length = 64 - int32(val&0b11_1111)
 	case 0xFF21:
-		a.ch4.envelope.initialVolume = int(val>>4) & 0b1111
+		a.ch4.envelope.initialVolume = int32(val>>4) & 0b1111
 		a.ch4.envelope.direction = util.Bit(val, 3)
-		a.ch4.envelope.speed = int(val & 0b111)
+		a.ch4.envelope.speed = int32(val & 0b111)
 		if !a.ch4.dacEnable() {
 			a.ch4.enabled = false
 		}
 	case 0xFF22:
-		a.ch4.octave = int(val >> 4) // ノイズ周波数2(オクターブ指定)
-		a.ch4.divisor = int(val & 0b111)
+		a.ch4.octave = int32(val >> 4) // ノイズ周波数2(オクターブ指定)
+		a.ch4.divisor = int32(val & 0b111)
 		a.ch4.period = a.ch4.calcFreqency()
 		a.ch4.width = 15
 		if util.Bit(val, 3) {
@@ -153,8 +153,8 @@ func (a *apu) Write(addr uint16, val uint8) {
 		}
 
 	case 0xFF24:
-		a.volume[0] = int(val>>4) & 0b111
-		a.volume[1] = int(val & 0b111)
+		a.volume[0] = (val >> 4) & 0b111
+		a.volume[1] = (val & 0b111)
 
 	case 0xFF25:
 		a.ch1.ignored = !util.Bit(val, 0)
